requirements.txt
pythonocc-core>=7.7.0
